+++ base_aux/breeders/test__breeder_2_objects.py
@@ -26,7 +26,7 @@ class ItemList(ItemSingle):
 class Test__BreederObjectList:
     # -----------------------------------------------------------------------------------------------------------------
     def test__groups_check__generated(self):
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             CLS_SINGLE__ITEM_SINGLE = ItemSingle
             CLS_LIST__ITEM_LIST = ItemList
 
@@ -34,7 +34,7 @@ class Test__BreederObjectList:
         Victim.generate__objects()
         assert Victim.groups_check__generated() is True
 
-        class Victim2(BreederObjectList):
+        class Victim2(BreederObj):
             CLS_SINGLE__ITEM_SINGLE = ItemSingle
             CLS_LIST__ITEM_LIST = ItemList
 
@@ -45,7 +45,7 @@ class Test__BreederObjectList:
     # -----------------------------------------------------------------------------------------------------------------
     def test__count(self):
         for count in range(5):
-            class Victim(BreederObjectList):
+            class Victim(BreederObj):
                 COUNT = count
                 CLS_SINGLE__ITEM_SINGLE = ItemSingle
                 CLS_LIST__ITEM_LIST = ItemList
@@ -63,26 +63,26 @@ class Test__BreederObjectList:
 
     # -----------------------------------------------------------------------------------------------------------------
     def test__groups__get_names(self):
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             pass
             # CLS_SINGLE__ITEM_SINGLE = ItemSingle
             # CLS_LIST__ITEM_LIST = ItemList
 
         assert Victim.groups__get_names() == list()
 
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             # CLS_SINGLE__ITEM_SINGLE = ItemSingle
             CLS_LIST__ITEM_LIST = ItemList
 
         assert Victim.groups__get_names() == ["ITEM_LIST", ]
 
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             CLS_SINGLE__ITEM_SINGLE = ItemSingle
             # CLS_LIST__ITEM_LIST = ItemList
 
         assert Victim.groups__get_names() == ["ITEM_SINGLE", ]
 
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             CLS_SINGLE__ITEM_SINGLE = ItemSingle
             CLS_LIST__ITEM_LIST = ItemList
 
@@ -90,7 +90,7 @@ class Test__BreederObjectList:
 
     # -----------------------------------------------------------------------------------------------------------------
     def test__groups_count__generated(self):
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             pass
             # CLS_SINGLE__ITEM_SINGLE = ItemSingle
             # CLS_LIST__ITEM_LIST = ItemList
@@ -99,7 +99,7 @@ class Test__BreederObjectList:
         Victim.generate__objects()
         assert Victim.groups_count__generated() == 0
 
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             # CLS_SINGLE__ITEM_SINGLE = ItemSingle
             CLS_LIST__ITEM_LIST = ItemList
 
@@ -107,7 +107,7 @@ class Test__BreederObjectList:
         Victim.generate__objects()
         assert Victim.groups_count__generated() == 1
 
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             CLS_SINGLE__ITEM_SINGLE = ItemSingle
             # CLS_LIST__ITEM_LIST = ItemList
 
@@ -115,7 +115,7 @@ class Test__BreederObjectList:
         Victim.generate__objects()
         assert Victim.groups_count__generated() == 1
 
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             CLS_SINGLE__ITEM_SINGLE = ItemSingle
             CLS_LIST__ITEM_LIST = ItemList
 
@@ -125,7 +125,7 @@ class Test__BreederObjectList:
 
     # -----------------------------------------------------------------------------------------------------------------
     def test__group_get__type(self):
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             COUNT = 2
             CLS_SINGLE__ITEM_SINGLE = ItemSingle
             CLS_LIST__ITEM_LIST = ItemList
@@ -137,7 +137,7 @@ class Test__BreederObjectList:
 
     # -----------------------------------------------------------------------------------------------------------------
     def test__group_check__exists(self):
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             pass
             # CLS_SINGLE__ITEM_SINGLE = ItemSingle
             # CLS_LIST__ITEM_LIST = ItemList
@@ -145,21 +145,21 @@ class Test__BreederObjectList:
         assert Victim.group_check__exists("ITEM_SINGLE") is False
         assert Victim.group_check__exists("ITEM_LIST") is False
 
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             # CLS_SINGLE__ITEM_SINGLE = ItemSingle
             CLS_LIST__ITEM_LIST = ItemList
 
         assert Victim.group_check__exists("ITEM_SINGLE") is False
         assert Victim.group_check__exists("ITEM_LIST") is True
 
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             CLS_SINGLE__ITEM_SINGLE = ItemSingle
             # CLS_LIST__ITEM_LIST = ItemList
 
         assert Victim.group_check__exists("ITEM_SINGLE") is True
         assert Victim.group_check__exists("ITEM_LIST") is False
 
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             CLS_SINGLE__ITEM_SINGLE = ItemSingle
             CLS_LIST__ITEM_LIST = ItemList
 
@@ -168,7 +168,7 @@ class Test__BreederObjectList:
 
     # -----------------------------------------------------------------------------------------------------------------
     def test__group_get__cls(self):
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             pass
             # CLS_SINGLE__ITEM_SINGLE = ItemSingle
             # CLS_LIST__ITEM_LIST = ItemList
@@ -176,21 +176,21 @@ class Test__BreederObjectList:
         assert Victim.group_get__cls("ITEM_SINGLE") is None
         assert Victim.group_get__cls("ITEM_LIST") is None
 
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             # CLS_SINGLE__ITEM_SINGLE = ItemSingle
             CLS_LIST__ITEM_LIST = ItemList
 
         assert Victim.group_get__cls("ITEM_SINGLE") is None
         assert Victim.group_get__cls("ITEM_LIST") is ItemList
 
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             CLS_SINGLE__ITEM_SINGLE = ItemSingle
             # CLS_LIST__ITEM_LIST = ItemList
 
         assert Victim.group_get__cls("ITEM_SINGLE") is ItemSingle
         assert Victim.group_get__cls("ITEM_LIST") is None
 
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             CLS_SINGLE__ITEM_SINGLE = ItemSingle
             CLS_LIST__ITEM_LIST = ItemList
 
@@ -199,7 +199,7 @@ class Test__BreederObjectList:
 
     # -----------------------------------------------------------------------------------------------------------------
     def test__group_get__insts(self):
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             pass
             # CLS_SINGLE__ITEM_SINGLE = ItemSingle
             # CLS_LIST__ITEM_LIST = ItemList
@@ -211,7 +211,7 @@ class Test__BreederObjectList:
         # assert isinstance(Victim.group_get__insts("ITEM_LIST"), list)
         # assert isinstance(Victim.group_get__insts("ITEM_LIST")[0], ItemList)
 
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             # CLS_SINGLE__ITEM_SINGLE = ItemSingle
             CLS_LIST__ITEM_LIST = ItemList
 
@@ -227,7 +227,7 @@ class Test__BreederObjectList:
         assert isinstance(Victim.group_get__insts("ITEM_LIST")[0], ItemList)
         assert len(Victim.group_get__insts("ITEM_LIST")) == Victim.COUNT
 
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             CLS_SINGLE__ITEM_SINGLE = ItemSingle
             # CLS_LIST__ITEM_LIST = ItemList
 
@@ -243,7 +243,7 @@ class Test__BreederObjectList:
         # assert isinstance(Victim.group_get__insts("ITEM_LIST")[0], ItemList)
         # assert len(Victim.group_get__insts("ITEM_LIST")) == Victim.COUNT
 
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             CLS_SINGLE__ITEM_SINGLE = ItemSingle
             CLS_LIST__ITEM_LIST = ItemList
 
@@ -261,7 +261,7 @@ class Test__BreederObjectList:
 
     # -----------------------------------------------------------------------------------------------------------------
     def test__group_call(self):
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             COUNT = 2
             CLS_SINGLE__ITEM_SINGLE = ItemSingle
             CLS_LIST__ITEM_LIST = ItemList
@@ -315,7 +315,7 @@ class Test__BreederObjectList:
 
     # -----------------------------------------------------------------------------------------------------------------
     def test__breeder_cls__and_getattr(self):
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             COUNT = 2
             CLS_SINGLE__ITEM_SINGLE = ItemSingle
             CLS_LIST__ITEM_LIST = ItemList
@@ -351,7 +351,7 @@ class Test__BreederObjectList:
 
     # -----------------------------------------------------------------------------------------------------------------
     def test__breeder_inst__and_getattr(self):
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             COUNT = 3
             CLS_SINGLE__ITEM_SINGLE = ItemSingle
             CLS_LIST__ITEM_LIST = ItemList
@@ -367,7 +367,7 @@ class Test__BreederObjectList:
 
     # -----------------------------------------------------------------------------------------------------------------
     def test__breeder_inst__deep(self):
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             COUNT = 3
             CLS_SINGLE__ITEM_SINGLE = ItemSingle
             CLS_LIST__ITEM_LIST = ItemList
@@ -392,7 +392,7 @@ class Test__BreederObjectList:
 
     # -----------------------------------------------------------------------------------------------------------------
     def test__generate_objects(self):
-        class Victim(BreederObjectList):
+        class Victim(BreederObj):
             COUNT = 2
             CLS_SINGLE__ITEM_SINGLE = ItemSingle
             CLS_LIST__ITEM_LIST = ItemList